TST: Separated out UpdatingAtomGroup tests
